feat: feat/rework-golang
