+++ go/static.go
@@ -1,0 +1,25 @@
+package main
+
+import (
+	"context"
+	"dagger/go/internal/dagger"
+)
+
+// Lint runs the linter on the provided source code
+func (m *Go) Lint(
+	ctx context.Context,
+	src *dagger.Directory,
+	// +optional
+	// +default="500s"
+	timeout string,
+) *dagger.Container {
+
+	golangciLintRunOpts := dagger.GolangciLintRunOpts{
+		Timeout: timeout,
+	}
+
+	return dag.GolangciLint().Run(
+		src,
+		golangciLintRunOpts,
+	)
+}
