--- tests/calculator/.golangci.yml
@@ -1,12 +1,0 @@
-run:
-  files:
-    - main.go
-
-linters:
-  disable-all: true
-  enable:
-    - govet
-    - gosimple
-    - staticcheck
-    - unused
-    - errcheck
